Reward message and success handling rewrite
The logic behind rewards has been updated to resolve when we get
expected behaviour, regardless of whether the reward was successful or
not. The returned object from rewards contains a success value, and a
message to log (eventually may be used for chat)

When handling reward execution, we now check for handled rewards and
leave unhandled rewards alone. This allows us to now raise errors on
unhandled exceptions.