Skip rewards we've already handled
Twitch is increasing their lazy loading, as as we're monitoring all
subtree mutations there's a high change we'll encounter a parent node
containing the same reward.

This change introduces a list of rewards we've already handled, and
appends to it on first handle. We skip the reward id if we encounter it
again.