nooms events
## Events
created three events that manipulate scenes

## Websocket Lib
added websocket lib to assets. We can properly bundle these with rollup or something later

## Rewards update
changed reward structure to be objects with a method for executing the code, suffice and cooldown is now definted on the object. Rewards are now more like definitions than just functions

## Cooldowns
Script now tracks cooldowns for rewards and adds items to a cooldowns queue, items are rejected if they're on cooldown (refunded)